requirements.txt
fastapi==0.95.1
uvicorn==0.23.1
sqlalchemy==2.0.18
jinja2==3.1.2
python-multipart==0.0.6
psycopg2-binary==2.9.7
python-dotenv==1.0.0